+++ autodocodec-api-usage/test_resources/nix/expression/these-object.nix
@@ -1,0 +1,5 @@
+{
+  int = 5;
+  text = "a";
+  type = "both";
+}