+++ autodocodec-nix/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.1.0.0] - 2025-02-13
+
+### Added
+
+* The `Autodocodec.Nix.Render` module for rendering values as Nix expressions.
+
+### Changed
+
+* Moved the options code into `Autodocodec.Nix.Options`.
+* Moved the expressions code into `Autodocodec.Nix.Expression`.
+
 ## [0.0.1.5] - 2024-11-04
 
 ### Changed